+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Register commands only for odata language
+- Property names in output from diagnostics fixed
 
 ## [0.1.5] - 2025-04-06